profile register and login tested but cookies doesn´t working

+++ Site_Libreria_G12/src/controllers/userController.js
@@ -3,6 +3,8 @@ const {
 } = require('express-validator');
 const bcryptjs = require('bcryptjs');
 const db = require ('../database/models');
+const fs = require('fs')
+const path = require('path')
 module.exports = {
   register: (req, res) => {
       res.render('register')
@@ -42,6 +44,14 @@ module.exports = {
       res.render('login')
   },
   loginUser: (req, res) => {
+
+    /* 
+    ------  Para acceder como administrador ----
+
+      usuario : noeliaromina20@gmail.com
+      contraseña: 123456
+    
+    */
       const errors = validationResult(req); 
 
       if (errors.isEmpty()) {
@@ -59,16 +69,17 @@ module.exports = {
               rolId : user.rolId
           }
 
-       // levanto session
-          if (req.body.remember) {
+       
+         /*  if (req.body.checkbox) {
               res.cookie('boulevardCookie', req.session.userLogin, {
                   maxAge: 1000 * 60 * 2
               })
-          }
+          } */
           res.redirect('/') 
       }).catch (error => console.log(error)) 
       } else {
           res.render('login', {
+            old: req.body,
               errors: errors.mapped()
           })
 
@@ -92,15 +103,34 @@ module.exports = {
         }))
         .catch (error => console.log(error)) 
 },
+
+/* 
+        IMPORTANTE : 
+
+      ----  Error al actualizar el perfil de usuario múltiples veces -----
+
+      al momento de actualizar los datos, todo funciona con normalidad y se actualiza el user en la db
+
+      pero si en la misma sesión quiero editar nuevamente el usuario, al mandar el formulario me devuelve un 404
+
+      Llevo dos dias tratando de solucionarlo pero no pude. La única manera de que funcione es bajar la sesión y volver a levantarla
+      
+      No me puedo guiar con proyectos ajenos ya que mi compañero armó la vista de edicion separada de la de perfil, entonces traté
+
+      de unir ambos archivos en uno solo y que se acceda a los datos a la vez que se puedan editar pero no actualizaba en la db
+
+
+
+*/
   update: (req, res) => {
     const errors = validationResult(req);
 
     if (errors.isEmpty()) {
-     /*  if(req.session.userLogin.image&&req.file&&req.session.userLogin.image!="noimage.jpeg"){
+      if(req.session.userLogin.image&&req.file&&req.session.userLogin.image!="noimage.jpeg"){
         fs.unlinkSync(
             path.resolve(__dirname,"..", "..", "public", "images",req.session.userLogin.image)
          )
-    } */
+    }
     let { name, lastname,birth,address,state,phone } = req.body;
     db.User.findByPk(req.session.userLogin.id)
       .then(() => {
@@ -109,7 +139,7 @@ module.exports = {
             name: name.trim(),
             lastname: lastname.trim(),
               birth,
-              address : address.trim(),
+              address,
               state,
               phone,
             image: req.file && req.file.filename || "noimage.jpeg",
@@ -128,7 +158,7 @@ module.exports = {
       .catch(error => console.log(error))
      
     } else {
-      return res.render("profile", {
+      return res.render("profileEdit", {
         user: req.body,
         errors: errors.mapped(),
       });
